Extract alerts filter into own class
